Move signup bottom sheet dialog to login library
