+++ WordPress/src/main/java/org/wordpress/android/ui/accounts/LoginActivity.java
@@ -35,13 +35,13 @@ import org.wordpress.android.login.LoginMagicLinkSentFragment;
 import org.wordpress.android.login.LoginMode;
 import org.wordpress.android.login.LoginSiteAddressFragment;
 import org.wordpress.android.login.LoginUsernamePasswordFragment;
+import org.wordpress.android.login.SignupBottomSheetDialog;
+import org.wordpress.android.login.SignupBottomSheetDialog.SignupSheetListener;
 import org.wordpress.android.ui.ActivityLauncher;
 import org.wordpress.android.ui.RequestCodes;
 import org.wordpress.android.ui.accounts.SmartLockHelper.Callback;
 import org.wordpress.android.ui.accounts.login.LoginPrologueFragment;
 import org.wordpress.android.ui.accounts.login.LoginPrologueListener;
-import org.wordpress.android.ui.accounts.signup.SignupBottomSheetDialog;
-import org.wordpress.android.ui.accounts.signup.SignupBottomSheetDialog.SignupSheetListener;
 import org.wordpress.android.ui.accounts.signup.SignupEmailFragment;
 import org.wordpress.android.ui.accounts.signup.SignupGoogleFragment;
 import org.wordpress.android.ui.accounts.signup.SignupMagicLinkFragment;
