+++ libs/login/WordPressLoginFlow/src/main/res/layout-land/signup_bottom_sheet_dialog.xml
