+++ libs/login/WordPressLoginFlow/src/main/res/layout/signup_bottom_sheet_dialog.xml
